move register components
