+++ packages/hono-takibi/src/generators/zod-openapi-hono/app/register-component/generate-register-component.test.ts
