+++ packages/hono-takibi/src/generators/zod-openapi-hono/app/register-component/generate-register-component.ts
@@ -1,4 +1,4 @@
-import type { SecuritySchemes } from '../../../types'
+import type { SecuritySchemes } from '../../../../types'
 
 export function generateRegisterComponent(securitySchemes: SecuritySchemes) {
   return Object.entries(securitySchemes)
